[Feature] RTT Control Buffer and Status (#52)
* Feature: RTT Control for Buffer Description and Status Information

Adds support via `rtt_get_status()` and `rtt_get_buf_descriptor()` to enable
querying for RTT status and buffer information respectively.  Both features
contribute to and resolve #48.

Adds an optional `block_address` parameter to `rtt_start()` to enable
configuring the address for the RTT control block to support devices where the
control block cannot be found automatically.  This resolves #48.

* Support: Functional test for RTT Support

Adds a functional test for RTT support.  Firmware image is built using the RTT
library and echoes data over RTT that it reads over RTT.  Test checks that we
are able to send and receive the echo back on the host.