Library: do not assume Linux implies GNU libc
On Linux, we rely on the dlinfo(3) API to retrieve
the JLink DLL's absolute file path corresponding to
the soname returned by the ctypes function find_library().

Unfortunately dlinfo() is not a POSIX API, but a GNU extension,
only available when using the GNU libc implementation.

This patch limits the dlinfo() dance to Linux systems
with GNU libc (glibc).