[Dependencies] Remove dependency on future
Currently the python-future package is only used to provide `input`.
However, `six` is also used within the source code for python2
compatibility. Replace the usage of `future` with `six` to trim the
required dependencies.

python-future also does not provide built wheels, so removing it means
that this package and all it's dependencies can be fetched as wheels
without running any local build steps.