[PR-182][Bug] Loading library on Aarch64 fails because pylink attempts to load 32-bit library (#182)
The ARM64 version of the JLink software contains two library files, `libjlinkarm.so` (64 bit)
and `libjlinkarm_arm.so` (32 bit). Pylink attempts to load `libjlinkarm_arm.so` first, which fails
on ARM64 due to the architecture mismatch and causes an error. This change simply
performs a 'test load' of each library file found, skipping if it fails, fixing this issue on ARM64.